+++ test/e2e/conftest.py
@@ -6,7 +6,7 @@ from run import generate_app
 
 from voicevox_engine.preset import PresetManager
 from voicevox_engine.setting import SettingLoader
-from voicevox_engine.synthesis_engine import make_synthesis_engines
+from voicevox_engine.tts_pipeline import make_synthesis_engines
 from voicevox_engine.utility.core_version_utility import get_latest_core_version
