+++ test/test_synthesis_engine.py
@@ -9,10 +9,10 @@ import numpy
 
 from voicevox_engine.acoustic_feature_extractor import OjtPhoneme
 from voicevox_engine.model import AccentPhrase, AudioQuery, Mora
-from voicevox_engine.synthesis_engine import SynthesisEngine
+from voicevox_engine.tts_pipeline import SynthesisEngine
 
 # TODO: import from voicevox_engine.synthesis_engine.mora
-from voicevox_engine.synthesis_engine.synthesis_engine import (
+from voicevox_engine.tts_pipeline.synthesis_engine import (
     apply_intonation_scale,
     apply_output_sampling_rate,
     apply_output_stereo,
